WCSim/WCSim/develop merged into nuPRISM/WCSim/nuPRISM/develop

All merge conflicts resolved
But has *NOT* been built yet

Conflicts:
	.gitignore
	CMakeLists.txt
	GNUmakefile_full
	GNUmakefile_root
	README.md
	WCSim.cc
	WCSim.mac
	include/RnModel_Fit_Params.hh
	include/WCSimDetectorConstruction.hh
	include/WCSimDetectorMessenger.hh
	include/WCSimEnumerations.hh
	include/WCSimEventAction.hh
	include/WCSimGenerator_Radioactivity.hh
	include/WCSimPMTObject.hh
	include/WCSimPrimaryGeneratorAction.hh
	include/WCSimPrimaryGeneratorMessenger.hh
	include/WCSimRandomParameters.hh
	include/WCSimRootEvent.hh
	include/WCSimRootGeom.hh
	include/WCSimRootLinkDef.hh
	include/WCSimRunAction.hh
	include/WCSimSteppingAction.hh
	include/WCSimTrackingAction.hh
	include/WCSimWCAddDarkNoise.hh
	include/WCSimWCDigi.hh
	include/WCSimWCDigitizer.hh
	include/WCSimWCHit.hh
	include/WCSimWCPMT.hh
	include/WCSimWCSD.hh
	include/WCSimWCTrigger.hh
	include/jhfNtuple.h
	macros/visOGLQT.mac
	src/WCSimConstructCylinder.cc
	src/WCSimConstructGeometryTables.cc
	src/WCSimConstructMaterials.cc
	src/WCSimConstructPMT.cc
	src/WCSimDarkRateMessenger.cc
	src/WCSimDetectorConstruction.cc
	src/WCSimDetectorMessenger.cc
	src/WCSimEventAction.cc
	src/WCSimGenerator_Radioactivity.cc
	src/WCSimPMTObject.cc
	src/WCSimPMTQE.cc
	src/WCSimPhysicsListFactory.cc
	src/WCSimPrimaryGeneratorAction.cc
	src/WCSimPrimaryGeneratorMessenger.cc
	src/WCSimRootEvent.cc
	src/WCSimRootGeom.cc
	src/WCSimRootOptions.cc
	src/WCSimRunAction.cc
	src/WCSimStackingAction.cc
	src/WCSimTrackingAction.cc
	src/WCSimTuningMessenger.cc
	src/WCSimVisManager.cc
	src/WCSimWCAddDarkNoise.cc
	src/WCSimWCDAQMessenger.cc
	src/WCSimWCDigi.cc
	src/WCSimWCDigitizer.cc
	src/WCSimWCHit.cc
	src/WCSimWCPMT.cc
	src/WCSimWCSD.cc
	src/WCSimWCTrigger.cc
	verification-test-scripts/verification_HitsChargeTime.C

+++ include/WCSimWCAddDarkNoise.hh
@@ -18,7 +18,6 @@ class WCSimWCAddDarkNoise : public G4VDigitizerModule
 public:
   
   WCSimWCAddDarkNoise(G4String name, WCSimDetectorConstruction*, G4String);
-  //WCSimWCAddDarkNoise(G4String name, WCSimDetectorConstruction*, G4String);
   ~WCSimWCAddDarkNoise();
   
 public:
@@ -32,9 +31,9 @@ public:
   void SetConversion(double iconvrate){ ConvRate = iconvrate; }
   void SetDarkMode(int imode){DarkMode = imode;}
   int GetDarkMode(){return DarkMode;}
-  void SetDarkHigh(int idarkhigh){DarkHigh = idarkhigh;}
+  void SetDarkHigh(double idarkhigh){DarkHigh = idarkhigh;}
   double GetDarkHigh(){return DarkHigh;}
-  void SetDarkLow(int idarklow){DarkLow = idarklow;}
+  void SetDarkLow(double idarklow){DarkLow = idarklow;}
   double GetDarkLow(){return DarkLow;}
   void SetDarkWindow(int idarkwindow){DarkWindow = idarkwindow;}
   int GetDarkWindow(){return (int)(DarkWindow);}
@@ -55,7 +54,7 @@ private:
 
   WCSimDetectorConstruction* myDetector;
   G4String detectorElement;
-  
+
   std::vector<std::pair<double, double> > ranges;
   std::vector<std::pair<double, double> > result;
